+++ class/GaiaAlpha/Database.php
@@ -7,6 +7,8 @@ use PDOException;
 
 use GaiaAlpha\Env;
 
+use GaiaAlpha\Database\LoggedPDO;
+
 class Database
 {
     private ?PDO $pdo = null;
@@ -14,7 +16,7 @@ class Database
     public function __construct(string $dsn)
     {
         try {
-            $this->pdo = new PDO($dsn);
+            $this->pdo = new LoggedPDO($dsn);
             $this->pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
             $this->pdo->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
         } catch (PDOException $e) {